Lesson-21.-Authorizing-The-Question-Using-Policies

+++ app/Http/Controllers/QuestionsController.php
@@ -8,6 +8,11 @@ use App\Http\Requests\AskQuestionRequest;
 
 class QuestionsController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth')->except('index', 'show');
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -77,6 +82,8 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        $this->authorize("update", $question);
+
         return view('questions.edit', compact('question'));
     }
 
@@ -89,6 +96,8 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        $this->authorize("update", $question);
+
         $question->update($request->only('title', 'body'));
 
         return redirect()
@@ -104,6 +113,8 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        $this->authorize("delete", $question);
+
         $question->delete();
 
         return redirect()
